modularize:patientappointments page into managable componenets
